Refactor: Mapper클래스 구현
- DTO클래스에서 검증기능을 분리하기위해 만든 클래스.
- 해당 클래스에서는 DTO와 domain간의 변환을 시켜주는 역할을 가진 메서드들이 모여있음.
