+++ src/main/java/christmas/InputView.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class InputView {
     public VisitedDateDTO requestVisitedDate() {
         System.out.println("12월 중 식당 예상 방문 날짜는 언제인가요? (숫자만 입력해 주세요!)");
-        return new VisitedDateDTO(readLine());
+        return Mapper.toVisitedDateDTO(readLine());
     }
 
     public List<MenuAndAmount> requestOrders() {
